// src/app/pages/models/freightautofill.ts
export class freightAutoFill{
    transporter_id:any;
    vehicle_number:any;
    engine_number: any;
    chassis_number: any;
    owner_name: any;
    owner_address: any;
    owner_pan_number: any;
    driver_name: string;
    driver_father_name: string;
    driver_address: any;
    driver_contact: 9981848879;
    driver_license_number: any
}